requirements.txt
Flask
yt-dlp
ffmpeg-python
edge-tts
uuid
SpeechRecognition
requests
groq
python-dotenv
google-genai
